fedup: Parallelize package downloads with "async"
Normal "yum update" does this -- there's no reason fedup should not,
when it deals with even larger sets of packages. This could help the
"prepare to upgrade" stage quite a bit.

Testing without this patch shows single connection at a time behavior;
with patch, we fetch from multiple mirrors. (iftop.)

Note that the UI is a little bit broken -- we lose the progress bar.
Somehow yum manages to keep it in the face of async downloads, so I
suspect it's possible to fix. But I'd rather have parallel downloads
than a progress bar any day ;-).